3 bug fixed

+++ src/components/Snippets/tabs/Created.js
@@ -183,7 +183,7 @@ const Created = (data) => {
                     <div className='col mb-4' >
                     <CardCreate img={x.Imageurl} 
                     title={x.NFTName} count="401" subTitle={`<span>Highest bid</span> <span>1/1</span>`} 
-                    linkText={parseInt(x.NFTPrice/1000000)} dataall={x}
+                    linkText={parseFloat(x.NFTPrice/1000000)} dataall={x}
                     />
                     </div>                    
                     </>
