+++ src/components/Snippets/tabs/OnSale.js
@@ -223,7 +223,7 @@ const OnSale = (data) => {
                     <div className='col mb-4' >
                     <CardSale img={x.Imageurl} 
                     title={x.NFTName} count="4012" subTitle={`<span>Highest bid</span> <span>1/1</span>`} 
-                    linkText={parseInt(x.NFTPrice/1000000)} dataall={x}
+                    linkText={parseFloat(x.NFTPrice/1000000)} dataall={x}
                     />
                     </div>                    
                     </>
