+++ src/components/Snippets/tabs/Owned.js
@@ -219,7 +219,7 @@ const Owned = (data) => {
                     <div className='col mb-4' >
                     <CardOwn img={x.Imageurl} 
                     title={x.NFTName} count="401" subTitle={`<span>Highest bid</span> <span>1/1</span>`} 
-                    linkText={parseInt(x.NFTPrice/1000000)} dataall={x}
+                    linkText={parseFloat(x.NFTPrice/1000000)} dataall={x}
                     />
                     </div>                    
                     </>
